GRT-86 [REFACTOR] 아이콘 import 경로 index.ts로 일원화
