--- src/app/calendar/page.tsx
@@ -1,5 +1,0 @@
-const Page = () => {
-  return <div></div>;
-};
-
-export default Page;
